test jwt - wip

+++ src/main/java/gr/hua/dit/ds/springbootdemo/controller/UserController.java
@@ -28,7 +28,7 @@ public class UserController {
     @PostMapping("/saveUser")
     public String saveUser(@ModelAttribute User user, Model model){
         System.out.println("Roles: "+user.getRoles());
-        Integer id = userService.saveUser(user);
+        Long id = userService.saveUser(user);
         String message = "User '"+id+"' saved successfully !";
         model.addAttribute("msg", message);
         return "home";
